refactor(cast): inline Tempo helpers, remove `src/tempo.rs`
