+++ crates/cast/src/lib.rs
@@ -62,7 +62,6 @@ pub mod call_spec;
 pub(crate) mod debug;
 pub mod errors;
 mod rlp_converter;
-pub mod tempo;
 pub mod tx;
 
 use rlp_converter::Item;
